# server/src/http/routes/swagger.yml
openapi: 3.0.1
info:
  title: API Exposition InserJeunes
  description: |
    L’API Exposition InserJeunes permet **la récupération des [données InserJeunes](https://documentation.exposition.inserjeunes.beta.gouv.fr/)** (dont la “part en emploi à 6 mois” et le “taux en formation”).

    Les diverses routes permettent de récupérer ces données : 

    - pour une **formation / certification,** à partir d’un code MEFSTAT11, pour la voie scolaire, ou d’un code formation diplôme (CFD), pour la voie apprentissage.
    - pour une **formation / certification au sein d’un établissement,** à partir d’un code MEFSTAT11 ou CFD et d’un code UAI (Unité Administrative Immatriculée).

    L’API permet en outre de récupérer plusieurs de ces éléments à la fois, et propose divers filtres, par exemple pour récupérer les données à la maille régionale ou académique.

    Certaines routes, mentionnées par un cadenas, nécessitent une authentification. Pour obtenir une clef API, nous vous invitons à [contacter nos équipes](mailto:contact@inserjeunes.beta.gouv.fr).

    Le code source de l’API est public et disponible [sur Github](https://github.com/mission-apprentissage/trajectoires-pro/).

  license:
    name: MIT License
    url: https://raw.githubusercontent.com/mission-apprentissage/trajectoires-pro/main/LICENSE
  version: 1.0.0
paths:
...
  /api/inserjeunes/certifications.{ext}:
    get:
      description: Exporter des statistiques de certifications
      security:
        - BearerAuth: [ ]
        - ApiKeyAuth: [ ]
      parameters:
        - $ref: "#/components/parameters/millesimes"
        - $ref: "#/components/parameters/code_certifications"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/items_par_page"
        - $ref: "#/components/parameters/ext"
      responses:
        200:
          $ref: "#/components/responses/CertificationsResponse"
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        404:
          $ref: "#/components/responses/404"
...
  /api/inserjeunes/formations.{ext}:
    get:
      description: Exporter des statistiques de formations
      security:
        - BearerAuth: [ ]
        - ApiKeyAuth: [ ]
      parameters:
        - $ref: "#/components/parameters/ext"
        - $ref: "#/components/parameters/uais"
        - $ref: "#/components/parameters/regions"
        - $ref: "#/components/parameters/academies"
        - $ref: "#/components/parameters/millesimesFormation"
        - $ref: "#/components/parameters/code_certifications"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/items_par_page"
      responses:
        200:
          $ref: "#/components/responses/FormationsResponse"
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        404:
          $ref: "#/components/responses/404"
...
  /api/inserjeunes/formations/{uai}-{code_certification}/widget:
    get:
      description: Retourne le code à utiliser pour intégrer le widget
      security:
        - BearerAuth: [ ]
      parameters:
        - $ref: "#/components/parameters/uai"
        - $ref: "#/components/parameters/code_certification"
        - $ref: "#/components/parameters/millesimePeriode"
      responses:
        200:
          $ref: "#/components/responses/FormationWidgetResponse"
components:
  parameters:
...
    millesimes:
      name: millesimes
      in: query
      description: |
        Permet de filtrer la recherche avec un ou plusieurs millesimes séparés par des virgules.

        * Les millesimes des certifications sont des années (ex: 2020)
        * Retourne les données du dernier millésime si le champ est vide
      schema:
        type: string
...
    millesimesFormation:
      name: millesimes
      in: query
      description: |
        Permet de filtrer la recherche avec un ou plusieurs millesimes séparés par des virgules.
        * Les millesimes des formations sont des périodes (ex: 2020_2021)
        * Retourne les données du dernier millésime si le champ est vide
      schema:
        type: string
    code_certifications:
      name: code_certifications
      in: query
      description: |
        Permet de filtrer la recherche avec un ou plusieurs codes de certification séparés par des virgules ou des |.

        Un code certification peut-être :
          * un code formation diplome (CFD) pour la voie apprentissage
          * un code MEFSTAT11 pour la voie professionnelle
      schema:
        type: string
    uais:
      name: uais
      in: query
      description: Permet de filtrer la recherche avec une ou plusieurs UAI séparées par des virgules.
      schema:
        type: string
        pattern: "^[0-9]{7}[A-Z]{1}$"
    regions:
      name: regions
      in: query
      schema:
        type: string
      description: |
        Permet de filtrer la recherche avec un ou plusieurs codes de région séparés par des virgules.
        
        Valeurs possibles :
         ```
         84 Auvergne-Rhône-Alpes
         27 Bourgogne-Franche-Comté
         53 Bretagne
         24 Centre-Val de Loire
         00 Collectivités d'outre-mer
         94 Corse
         44 Grand Est
         01 Guadeloupe
         03 Guyane
         32 Hauts-de-France
         04 La Réunion
         02 Martinique
         06 Mayotte
         28 Normandie
         75 Nouvelle-Aquitaine
         76 Occitanie
         52 Pays de la Loire
         93 Provence-Alpes-Côte d'Azur
         11 Île-de-France
         ```
    academies:
      name: academies
      in: query
      schema:
        type: string
      description: |
        Permet de filtrer la recherche avec un ou plusieurs codes d'académie séparés par des virgules.
        
        Valeurs possibles :
        ```
        01 Paris                       02 Aix-Marseille
        03 Besançon                    04 Bordeaux
        06 Clermont-Ferrand            07 Dijon
        08 Grenoble                    09 Lille
        10 Lyon                        11 Montpellier
        12 Nancy-Metz                  13 Poitiers
        14 Rennes                      15 Strasbourg
        16 Toulouse                    17 Nantes
        18 Orléans-Tours               19 Reims
        20 Amiens                      22 Limoges
        23 Nice                        24 Créteil
        25 Versailles                  27 Corse
        28 La Réunion                  31 Martinique
        32 Guadeloupe                  33 Guyane
        40 Nouvelle-Calédonie          41 Polynésie Française
        42 Wallis et Futuna            43 Mayotte
        44 Saint-Pierre-et-Miquelon    70 Normandie
        77 Saint-Barthélemy            78 Saint-Martin
        ```
    page:
      name: page
      in: query
      description: Le numéro de la page désirée
      schema:
        type: integer
    items_par_page:
      name: items_par_page
      in: query
      description: Le nombre maximum d'éléments dans la page
      schema:
        type: integer
        default: 10
...
    millesimePeriode:
      name: millesime
      in: query
      description: |
        Les millesimes au niveau régional sont des périodes (ex: 2020_2021).

        Retourne les données du dernier millésime si le champ est vide.
      required: false
      schema:
        type: string
...
    uai:
      name: uai
      in: path
      required: true
      schema:
        type: string
        pattern: "^[0-9]{7}[A-Z]{1}$"
        example: "0060033D"
...
    code_certification:
      name: code_certification
      description: | 
       Un code certification peut-être :
          * un code formation diplome (CFD) pour la voie apprentissage
          * un code MEFSTAT11 pour la voie professionnelle
      in: path
      required: true
      schema:
        type: string
        example: "32221031409"
...
    ext:
      name: ext
      in: path
      required: true
      schema:
        type: string
        enum: [ "json", "csv" ]
        default: json
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        data:
          description: "Optionnel"
          type: object
          properties:
            millesimes:
              type: string
              description: "Millésime demandé (ou dernier millésime)"
            millesimesDisponible:
              type: array
              description: "Millésimes avec des données"
              items:
                type: string
    Pagination:
      type: object
      properties:
        page:
          type: string
        resultats_par_page:
          type: number
        nombre_de_page:
          type: number
        total:
          type: number
      description: Statut de la pagination
    Region:
      type: object
      required:
        - code
        - nom
      properties:
        code:
          type: string
          description: Le code INSEE de la région
        nom:
          type: string
          description: Le nom de la région
    Meta:
      type: object
      required:
        - date_import
      properties:
        date_import:
          type: string
        title:
          type: string
        description:
          type: string
        messages:
          description: Informations supplémentaires sur les données
          type: array
          items:
            type: string
          example: ["Les taux ne peuvent pas être affichés car il n'y a pas assez d'élèves pour fournir une information fiable."]
    DonneeSource:
      type: object
      required:
        - code
        - nom
      properties:
        code_certification:
          type: string
          description: Le code de la certification d'ou proviennent les données (Similaire dans le cas type = self)
        type:
          type: string
          description: |
            Le type de la source des données :
              * self : les données proviennent de la certification
              * ancienne : les données proviennent d'une certification plus ancienne
              * nouvelle : les données proviennent d'une certification plus récente
    Certification:
      type: object
      description: |
        Statistiques pour une certification.

        Les taux retournés sont :
          * null quand les effectifs (nb_annee_term) sont < 20
          * absent si l'on ne peut pas les calculer (données manquantes)
      required:
        - millesime
        - code_certification
        - filiere
        - libelle
        - _meta
      properties:
        millesime:
          type: string
        code_certification:
          type: string
        filiere:
          type: string
          enum: [ "apprentissage", "pro" ]
        libelle:
          description: "Libellé de la certification"
          type: string
        libelle_ancien:
          description: "Ancien libellé de la certification (disponible en cas de changement de nom)"
          type: string
        donnee_source:
          $ref: "#/components/schemas/DonneeSource"
        diplome:
          type: object
          properties:
            code:
              type: string
            libelle:
              type: string
        nb_sortant:
          type: number
        nb_annee_term:
          type: number
        nb_poursuite_etudes:
          type: number
        nb_en_emploi_24_mois:
          type: number
        nb_en_emploi_18_mois:
          type: number
        nb_en_emploi_12_mois:
          type: number
        nb_en_emploi_6_mois:
          type: number
        taux_en_emploi_24_mois:
          type: number
          nullable: true
        taux_en_emploi_18_mois:
          type: number
          nullable: true
        taux_en_emploi_12_mois:
          type: number
          nullable: true
        taux_en_emploi_6_mois:
          type: number
          nullable: true
        taux_en_formation:
          type: number
          nullable: true
        taux_autres_6_mois:
          type: number
          nullable: true
        taux_autres_12_mois:
          type: number
          nullable: true
        taux_autres_18_mois:
          type: number
          nullable: true
        taux_autres_24_mois:
          type: number
          nullable: true
        _meta:
          $ref: "#/components/schemas/Meta"
...
    Formation:
      type: object
      description: |
        Statistiques d'une certification pour une formation.

        Les taux retournés sont :
          * null quand les effectifs (nb_annee_term) sont < 20
          * absent si l'on ne peut pas les calculer (données manquantes)
      required:
        - uai
        - millesime
        - code_certification
        - filiere
        - libelle
        - region
        - _meta
      properties:
        uai:
          type: string
        millesime:
          type: string
        code_certification:
          type: string
        filiere:
          type: string
          enum: [ "apprentissage", "pro" ]
        libelle:
          description: "Libellé de la certification"
          type: string
        libelle_ancien:
          description: "Ancien libellé de la certification (disponible en cas de changement de nom)"
          type: string
        donnee_source:
          $ref: "#/components/schemas/DonneeSource"
        nb_sortant:
          type: number
        nb_annee_term:
          type: number
        nb_poursuite_etudes:
          type: number
        nb_en_emploi_24_mois:
          type: number
        nb_en_emploi_18_mois:
          type: number
        nb_en_emploi_12_mois:
          type: number
        nb_en_emploi_6_mois:
          type: number
        taux_en_emploi_24_mois:
          type: number
          nullable: true
        taux_en_emploi_18_mois:
          type: number
          nullable: true
        taux_en_emploi_12_mois:
          type: number
          nullable: true
        taux_en_emploi_6_mois:
          type: number
          nullable: true
        taux_en_formation:
          type: number
          nullable: true
        taux_autres_6_mois:
          type: number
          nullable: true
        taux_autres_12_mois:
          type: number
          nullable: true
        taux_autres_18_mois:
          type: number
          nullable: true
        taux_autres_24_mois:
          type: number
          nullable: true
        region:
          $ref: "#/components/schemas/Region"
        uai_type:
          type: string
          description: Typologie de l'UAI (lieu de formation, formateur, gestionnaire ou inconnu)
        uai_donnee:
          type: string
          description: UAI de provenance de la donnée
        uai_donnee_type:
          type: string
          description: Typologie de l'UAI (lieu de formation, formateur, gestionnaire ou inconnu)
        uai_lieu_formation:
          type: array
          description: Tableau des UAIs des lieux de formations associés à la formation dans le cas ou le type de l'UAI est formateur ou gestionnaire
          items:
            type: string
        uai_formateur:
          type: array
          description: Tableau des UAIs formateurs associés à la formation dans le cas ou le type de l'UAI est lieu de formation ou gestionnaire
          items:
            type: string
        uai_gestionnaire:
          type: array
          description: Tableau des UAIs gestionnaires associés à la formation dans le cas ou le type de l'UAI est lieu de formation ou formateur
          items:
            type: string
        _meta:
          $ref: "#/components/schemas/Meta"
...
  responses:
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    '401':
      description: L'api key est manquante ou invalide
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    CertificationsResponse:
      description: La liste des certifications avec leurs statistiques
      content:
        "application/json":
          schema:
            type: object
            properties:
              certifications:
                type: array
                items:
                  $ref: "#/components/schemas/Certification"
              pagination:
                $ref: "#/components/schemas/Pagination"
        "text/csv; charset=UTF-8":
          schema:
            type: string
            example: |
              code_certification;filiere;millesime;donnee_source_type;donnee_source_code_certification;nb_annee_term;nb_en_emploi_12_mois;nb_en_emploi_18_mois;nb_en_emploi_24_mois;nb_en_emploi_6_mois;nb_poursuite_etudes;nb_sortant;taux_autres_12_mois;taux_autres_18_mois;taux_autres_24_mois;taux_autres_6_mois;taux_en_emploi_12_mois;taux_en_emploi_18_mois;taux_en_emploi_24_mois;taux_en_emploi_6_mois;taux_en_formation;taux_rupture_contrats
              50034303;apprentissage;2019;self;50034303;9;3;3;3;4;0;9;67;67;67;56;33;33;33;44;0;20
...
    FormationsResponse:
      description: La liste des statistiques des formations
      content:
        "application/json":
          schema:
            type: object
            properties:
              formations:
                type: array
                items:
                  $ref: "#/components/schemas/Formation"
              pagination:
                $ref: "#/components/schemas/Pagination"
        "text/csv; charset=UTF-8":
          schema:
            type: string
            example: |
              uai;code_certification;filiere;millesime;donnee_source_type;donnee_source_code_certification;nb_annee_term;nb_en_emploi_12_mois;nb_en_emploi_18_mois;nb_en_emploi_24_mois;nb_en_emploi_6_mois;nb_poursuite_etudes;nb_sortant;taux_autres_12_mois;taux_autres_18_mois;taux_autres_24_mois;taux_autres_6_mois;taux_en_emploi_12_mois;taux_en_emploi_18_mois;taux_en_emploi_24_mois;taux_en_emploi_6_mois;taux_en_formation;taux_rupture_contrats
              0010826T;01022103;apprentissage;2019_2020;self;01022103;9;5;;;4;4;5;0;;;11;56;;;44;44;
...
    FormationWidgetResponse:
      description: Une iframe qui permet d'intégrer le widget
      content:
        "text/html":
          schema:
            type: string
          examples:
            html:
              summary: Une iframe permettant d'intégrer le widget
              value: |
                <iframe onLoad="JS_CODE" style="width: 100%; height: 0;" src="/api/inserjeunes/formations/[UAI]-[CODE_CERTIFICATION]/widget/[HASH]?" scrolling="no" frameBorder="0" />
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: query
      name: apiKey
    BearerAuth:
      type: http
      scheme: bearer